feat(CombineGWASWorkflowAttritionTableLogic): Reorganized files into AttritionTable directory
