+++ src/Analysis/GWASV2/Shared/AttritionTableWrapper/AttritionTableRow.jsx
@@ -9,7 +9,6 @@ import {
 } from '../wizardEndpoints/cohortMiddlewareApi';
 import BarChart from './ChartIcons/BarChart';
 import EulerDiagram from './ChartIcons/EulerDiagram';
-import './AttritionTable.css';
 
 const AttritionTableRow = ({
   otherCohortDefinitionId,
@@ -37,19 +36,20 @@ const AttritionTableRow = ({
     ],
     // if there are not two cohorts selected, then quantitative
     // Otherwise if there are two cohorts selected, case control
-    () => (!(cohortDefinitionId && otherCohortDefinitionId)
-      ? fetchConceptStatsByHareSubset(
-        cohortDefinitionId,
-        covariateSubset,
-        sourceId,
-      )
-      : fetchConceptStatsByHareSubsetCC(
-        cohortDefinitionId,
-        otherCohortDefinitionId,
-        covariateSubset,
-        sourceId,
-      )),
-    queryConfig,
+    () =>
+      !(cohortDefinitionId && otherCohortDefinitionId)
+        ? fetchConceptStatsByHareSubset(
+            cohortDefinitionId,
+            covariateSubset,
+            sourceId
+          )
+        : fetchConceptStatsByHareSubsetCC(
+            cohortDefinitionId,
+            otherCohortDefinitionId,
+            covariateSubset,
+            sourceId
+          ),
+    queryConfig
   );
 
   const { breakdown } = { breakdown: data?.concept_breakdown };
@@ -57,7 +57,7 @@ const AttritionTableRow = ({
   const getSizeByColumn = (hare) => {
     const hareIndex = breakdownColumns.findIndex(
       // eslint-disable-next-line camelcase
-      ({ concept_value }) => concept_value === hare,
+      ({ concept_value }) => concept_value === hare
     );
     return hareIndex > -1
       ? breakdownColumns[hareIndex].persons_in_cohort_with_value
@@ -68,13 +68,13 @@ const AttritionTableRow = ({
     if (breakdown?.length) {
       const filteredBreakdown = breakdown.filter(
         // eslint-disable-next-line camelcase
-        ({ concept_value }) => concept_value !== 'OTH',
+        ({ concept_value }) => concept_value !== 'OTH'
       );
       setBreakdownSize(
         filteredBreakdown.reduce(
           (acc, curr) => acc + curr.persons_in_cohort_with_value,
-          0,
-        ),
+          0
+        )
       );
       setBreakdownColumns(filteredBreakdown);
     } else {
